deletecomment

+++ templates/expandedpost.tpl
@@ -20,6 +20,6 @@
 			{/if}
 		{/foreach}
 		
-		<a href="deletepost.php?postid={$postid}">Delete post</a>
+		<a href="deletepost.php?postid={$postid}" onclick="return confirm('Are you sure you want to delete this post?');">Delete post</a>
 	</div>
 </body>
